Move DifferentiationInvoker to a separate file.
Part of https://bugs.swift.org/browse/TF-993
